requirements.txt
flask
pandas
numpy
matplotlib
seaborn
PYyaml
sklearn
xgboost
gunicorn
ipykernel
cassandra-driver
evidently
-e .